pyproject.toml
[build-system]
requires = ["setuptools >= 67.3.2", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "mob-tool"
version = "1.0.27"
authors = [
    { name = "rfst" },
]
description = "Mob session management tool"
readme = "README.md"
requires-python = ">=3.7"
keywords = [
    "mob", "mob-programming", "timer", "mob-programming-timer", "cli-tool",
    "remote-mob-programming", "remote-mobs", "mob-timer", "mob-tool",
]
license = { file = "LICENSE.txt" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python"
]
dependencies = [
    'click==8.1.3',
    'injector==0.20.1',
    'dataclasses-json==0.5.7',
    'GitPython==3.1.30',
    'colorama==0.4.6',
    'requests==2.28.2',
    'packaging==23.0',
    'flet==0.4.0',
    'platformdirs==3.0.0',
]

[project.scripts]
mobt = "mobt.Controllers:cli"
mob-tool = "mobt.Controllers:cli"

[project.urls]
source = "https://github.com/stavarengo/mob-tool"
Say_Thanks = "https://saythanks.io/to/faelsta"
